Split GoalView
