+++ frontend/src/views/goals/GoalView.js
@@ -1,32 +1,19 @@
 import React, { useState } from 'react'
-import {
-  Card, CardHeader, ListItemText, List, ListItem, ListItemIcon, IconButton, MenuItem, Menu
-} from '@material-ui/core'
-import {
-  ArrowRight as ArrowRightIcon, ArrowLeft as ArrowLeftIcon, Delete as DeleteIcon, Edit as EditIcon
-} from '@material-ui/icons'
+import { Menu, MenuItem } from '@material-ui/core'
 import { makeStyles } from '@material-ui/core/styles'
 import { useMutation, useQuery } from 'react-apollo-hooks'
 
 import { WORKSPACE_BY_ID } from '../../graphql/Query'
 import LoadingBar from '../../components/LoadingBar'
 import NotFoundView from '../error/NotFoundView'
-import CourseEditor from '../manager/CourseEditor'
-import ConceptEditor from '../manager/ConceptEditor'
-import ConceptToolTipContent from '../../components/ConceptTooltipContent'
-import {
-  CREATE_CONCEPT, UPDATE_CONCEPT, DELETE_CONCEPT,
-  DELETE_COURSE, CREATE_COURSE, UPDATE_COURSE,
-  CREATE_GOAL_LINK, DELETE_GOAL_LINK
-} from '../../graphql/Mutation'
+import { CREATE_GOAL_LINK, DELETE_GOAL_LINK } from '../../graphql/Mutation'
 import cache from '../../apollo/update'
-import { useConfirmDelete } from '../../dialogs/alert'
-import { noPropagation } from '../../lib/eventMiddleware'
 import { ConceptLink } from '../coursemapper/concept'
 import { backendToSelect } from '../../dialogs/tagSelectUtils'
+import { Courses } from './Courses'
+import { Goals } from './Goals'
 
-
-const useStyles = makeStyles(theme => ({
+export const useStyles = makeStyles(theme => ({
   root: {
     width: '100%',
     maxWidth: '1280px',
@@ -76,166 +63,6 @@ const useStyles = makeStyles(theme => ({
   }
 
 }))
-
-const CourseItem = ({ course, deleteCourse, setEditing, onClickCircle }) => {
-  const classes = useStyles()
-  const confirmDelete = useConfirmDelete()
-
-  const handleDelete = noPropagation(async () => {
-    if (await confirmDelete(`Are you sure you want to delete the course ${course.name}?`)) {
-      await deleteCourse(course.id)
-    }
-  })
-
-  return (
-    <ListItem divider key={course.id} className={classes.listItemContainer}>
-      <ListItemText>{course.name}</ListItemText>
-      <ListItemIcon>
-        <IconButton onClick={handleDelete} className={classes.hoverButton}>
-          <DeleteIcon />
-        </IconButton>
-        <IconButton onClick={() => setEditing(course.id)} className={classes.hoverButton}>
-          <EditIcon />
-        </IconButton>
-        <IconButton
-          onClick={onClickCircle('course', course.id)}
-          className={classes.activeCircle}
-        >
-          <ArrowRightIcon
-            viewBox='7 7 10 10' id={`course-circle-${course.id}`}
-            className='course-circle' />
-        </IconButton>
-      </ListItemIcon>
-    </ListItem>
-  )
-}
-
-const GoalItem = ({ goal, deleteConcept, setEditing, onClickCircle }) => {
-  const classes = useStyles()
-  const confirmDelete = useConfirmDelete()
-
-  const handleDelete = noPropagation(async () => {
-    if (await confirmDelete(`Are you sure you want to delete the goal ${goal.name}?`)) {
-      await deleteConcept(goal.id)
-    }
-  })
-
-  return (
-    <ListItem divider key={goal.id} className={classes.listItemContainer}>
-      <ListItemIcon>
-        <IconButton
-          onClick={onClickCircle('goal', goal.id)} className={classes.activeCircle}
-        >
-          <ArrowLeftIcon
-            viewBox='7 7 10 10' id={`goal-circle-${goal.id}`}
-            className='goal-circle' />
-        </IconButton>
-      </ListItemIcon>
-      <ListItemText>
-        <ConceptToolTipContent
-          tags={goal.tags} subtitle={goal.description} description={goal.name}
-        />
-      </ListItemText>
-      <ListItemIcon>
-        <IconButton onClick={handleDelete} className={classes.hoverButton}>
-          <DeleteIcon />
-        </IconButton>
-        <IconButton onClick={() => setEditing(goal.id)} className={classes.hoverButton}>
-          <EditIcon />
-        </IconButton>
-      </ListItemIcon>
-    </ListItem>
-  )
-}
-
-const Goals = ({ workspaceId, goals, tagOptions, onClickCircle }) => {
-  const classes = useStyles()
-  const [editing, setEditing] = useState()
-
-  const createConcept = useMutation(CREATE_CONCEPT, {
-    update: cache.createConceptUpdate(workspaceId)
-  })
-  const updateConcept = useMutation(UPDATE_CONCEPT, {
-    update: cache.updateConceptUpdate(workspaceId)
-  })
-  const deleteConcept = useMutation(DELETE_CONCEPT, {
-    update: cache.deleteConceptUpdate(workspaceId)
-  })
-
-  return (
-    <Card elevation={0} className={classes.card}>
-      <CardHeader title='Goals' />
-      <List className={classes.list}>
-        {goals.map(goal => editing === goal.id ? (
-          <ConceptEditor
-            submit={args => {
-              setEditing(null)
-              return updateConcept({ variables: args })
-            }}
-            tagOptions={tagOptions}
-            cancel={() => setEditing(null)}
-            defaultValues={goal}
-            action='Save'
-            key={goal.id}
-          />
-        ) : (
-          <GoalItem
-            updateConcept={variables => updateConcept({ variables })}
-            deleteConcept={id => deleteConcept({ variables: { id } })}
-            key={goal.id} goal={goal} setEditing={setEditing} onClickCircle={onClickCircle}
-          />
-        ))}
-      </List>
-      <ConceptEditor submit={args => createConcept({
-        variables: {
-          workspaceId,
-          ...args
-        }
-      })} action='Create' tagOptions={tagOptions} defaultValues={{ level: 'GOAL' }} />
-    </Card>
-  )
-}
-
-const Courses = ({ workspaceId, courses, tagOptions, onClickCircle }) => {
-  const classes = useStyles()
-  const [editing, setEditing] = useState()
-
-  const createCourse = useMutation(CREATE_COURSE, { update: cache.createCourseUpdate(workspaceId) })
-  const updateCourse = useMutation(UPDATE_COURSE, { update: cache.updateCourseUpdate(workspaceId) })
-  const deleteCourse = useMutation(DELETE_COURSE, { update: cache.deleteCourseUpdate(workspaceId) })
-
-  return (
-    <Card elevation={0} className={classes.card}>
-      <CardHeader title='Courses' />
-      <List className={classes.list}>
-        {courses.map(course => editing === course.id ? (
-          <CourseEditor
-            submit={args => {
-              setEditing(null)
-              return updateCourse({ variables: args })
-            }}
-            tagOptions={tagOptions}
-            cancel={() => setEditing(null)}
-            defaultValues={course}
-            action='Save'
-            key={course.id}
-          />
-        ) : (
-          <CourseItem
-            deleteCourse={id => deleteCourse({ variables: { id } })}
-            key={course.id} course={course} setEditing={setEditing} onClickCircle={onClickCircle}
-          />
-        ))}
-      </List>
-      <CourseEditor tagOptions={tagOptions} submit={args => createCourse({
-        variables: {
-          workspaceId,
-          ...args
-        }
-      })} />
-    </Card>
-  )
-}
 
 const GoalView = ({ workspaceId }) => {
   const classes = useStyles()
